feature: main-page second.report 알고리즘 구현
