+++ relanz/main/views/__init__.py
@@ -1,0 +1,2 @@
+from .home_views import *
+from .ranking_views import *
